Split code (footer).
